requirements.txt
SQLAlchemy==2.0.44
tabulate==0.9.0
python-dateutil==2.8.2
Flask==2.3.2